🚀 feat(afloatApi.js): add methods to set afloat balance and add afloat admin
The `setAfloatBalance` method is added to set the balance of an address in the afloat pallet. It takes an `address` and `amount` as parameters and calls the `setAfloatBalance` extrinsic.

The `addAfloatAdmin` method is added to add an address as an admin in the afloat pallet. It takes an `address` as a parameter and calls the `addAfloatAdmin` extrinsic.